Extract the CLI implementation into an independent `boxlite-cli` crate to enable separate versioning, independent dependency management, and cleaner architecture.
