+++ boxlite-cli/src/main.rs
@@ -1,0 +1,24 @@
+mod cli;
+mod commands;
+
+use clap::Parser;
+use cli::Cli;
+
+#[tokio::main]
+async fn main() -> anyhow::Result<()> {
+    let cli = Cli::parse();
+
+    // Initialize logging based on global flags
+    if cli.global.debug {
+        // SAFETY: We set this early in main, racing with other threads reading env is unlikely/acceptable for CLI tool
+        unsafe {
+            std::env::set_var("RUST_LOG", "debug");
+        }
+    }
+
+    match cli.command {
+        cli::Commands::Run(args) => commands::run::execute(args).await?,
+    }
+
+    Ok(())
+}
