--- boxlite/src/cli/commands/image/mod.rs
@@ -1,1 +1,0 @@
-// Future image commands
