--- boxlite/src/cli/commands/mod.rs
@@ -1,2 +1,0 @@
-pub mod container;
-pub mod image;
